# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Service serving data from a CSV file to a REST API",
      "version": "1.0.0",
      "title": "Data Analyzer",
      "contact": {
        "email": "cg@jmg-conseil.eu"
      }
    },
    "servers": [
        {
          "url": "https://lab.jmg-conseil.eu",
          "description": "Development server"
        }
    ],
    "paths": {
      "/cell/docs": {
          "get": {
              "tags": [
                  "DOCS"
              ],
              "description": "OpenAPI 3.0 documentation",
              "responses": {
                  "200": {
                      "description": "Success"
                  },
                  "401": {
                      "description": "Unauthorized"
                  },
                  "403": {
                      "description": "Forbidden"
                  }
              }
          }
      },
      "/cell": {
        "get": {
            "tags": [
                "GET"
            ],
            "description": "List data for all stations.",
            "responses": {
                "200": {
                    "description": "Success: return results",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "403": {
                    "description": "Forbidden"
                }
            }
        }
      },
      "/cell/csv": {
        "get": {
            "tags": [
                "GET"
            ],
            "description": "Base CSV document",
            "responses": {
                "200": {
                    "description": "Success: return document",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "403": {
                    "description": "Forbidden"
                }
            }
        }
      },
      "/cell/station{{uic}}": {
        "get": {
            "tags": [
                "GET"
            ],
            "description": "Display all stations matching query (ex : https://lab.jmg-conseil.eu/cell/station?uic=0087784561&uic=0087781005)",
            "parameters": [
                {
                    "name": "uic",
                    "in": "query",
                    "description": "Requested uic-code. Several can be provided.",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Success",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "404": {
                    "description": "Zipcode not found"
                },
                "403": {
                    "description": "Forbidden"
                }
            }
        }
      }
    }
  }